refactored endpoint to api
