+++ config/default.json
@@ -7,7 +7,7 @@
     "SCALE": 2
   },
   "SIDECAR": {
-    "DISABLED": true,
+    "DISABLED": false,
     "HOST": "localhost",
     "PORT": 5678,
     "CONNECT_TIMEOUT": 45000,
