+++ src/api/metrics/plugin.js
